// apps/api-server/src/shared/domain/notification-type.enum.ts
export enum NotificationType {
  CALENDAR_EVENT = 'calendar_event',
  ACCOUNTABILITY_BUDDY_INVITATION = 'accountability_buddy_invitation',
  ACCOUNTABILITY_BUDDY_INVITATION_ACCEPTED = 'accountability_buddy_invitation_accepted',
  UNLOCK_REQUEST_RECEIVED = 'unlock_request_received',
  UNLOCK_REQUEST_APPROVED = 'unlock_request_approved',
  UNLOCK_REQUEST_REJECTED = 'unlock_request_rejected',
}